chore: refactor local authority comparator selection into a seperate controller
