+++ web/src/Web.App/Views/LocalAuthority/_HighNeeds.cshtml
@@ -35,7 +35,7 @@
             <feature negate="true" name="@FeatureFlags.HighNeedsBenchmarking">
                 <li>
                     <h3 class="govuk-heading-s">
-                        <a href="@Url.Action("Comparators", "LocalAuthorityHighNeedsBenchmarking", new
+                        <a href="@Url.Action("Index", "LocalAuthorityComparators", new
                                  {
                                      code = Model.Code
                                  })" class="govuk-link govuk-link--no-visited-state">
