+++ web/src/Web.App/Views/LocalAuthorityHighNeedsBenchmarking/Index.cshtml
@@ -16,7 +16,7 @@
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
         <p class="govuk-body">
-            <a href="@Url.Action("Comparators", "LocalAuthorityHighNeedsBenchmarking", new
+            <a href="@Url.Action("Index", "LocalAuthorityComparators", new
                      {
                          code = Model.Code,
                          referrer = "benchmarking"
